position properties were changed a type from int to float

+++ Domain.Core/Models/Pet.cs
@@ -33,20 +33,20 @@ namespace Domain.Core.Models
         public long DeathDate { get; set; } //in seconds
 
         public Guid BodyId { get; set; }
-        public int BodyPictureX { get; set; }
-        public int BodyPictureY { get; set; }
+        public float BodyPictureX { get; set; }
+        public float BodyPictureY { get; set; }
         public float BodyPictureScale { get; set; }
         public Guid EyesId { get; set; }
-        public int EyesPictureX { get; set; }
-        public int EyesPictureY { get; set; }
+        public float EyesPictureX { get; set; }
+        public float EyesPictureY { get; set; }
         public float EyesPictureScale { get; set; }
         public Guid NoseId { get; set; }
-        public int NosePictureX { get; set; }
-        public int NosePictureY { get; set; }
+        public float NosePictureX { get; set; }
+        public float NosePictureY { get; set; }
         public float NosePictureScale { get; set; }
         public Guid MouthId { get; set; }
-        public int MouthPictureX { get; set; }
-        public int MouthPictureY { get; set; }
+        public float MouthPictureX { get; set; }
+        public float MouthPictureY { get; set; }
         public float MouthPictureScale { get; set; }
 
         public Farm Farm { get; set; }
